chore(xrp): dedupe shared pubkey helper and clean e2e timeouts
Move compressSecp256k1Pubkey to one xrp shared module and re-export from chain/devices.

Replace repeated XRP e2e timeout literals with constants and localize one-off test client setup.
